+++ src/main/java/PruebaTecnica/RetoSofka/Repository/TiqueteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TiqueteRepository extends JpaRepository<Tiquete, Integer> {
-    List<Tiquete> findAllByCliente_Id(Long clienteId);
+    List<Tiquete> findAllByCliente_Id(Integer clienteId);
 
 }
